My C language commit

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char z;
 	float f;
 
-	printf("Size of a char: %d bytes(s)\n", (unsigned long)sizeof(z));
-	printf("Size of an int: %d bytes(s)\n", (unsigned long)sizeof(w));
-	printf("Size of a long int: %d bytes(s)\n", (unsigned long)sizeof(x));
-	printf("Size of a long long int: %d bytes(s)\n", (unsigned long)sizeof(y));
-	printf("Size of a float: %d bytes(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %d bytes(s)\n", sizeof(z));
+	printf("Size of an int: %d bytes(s)\n", sizeof(w));
+	printf("Size of a long int: %d bytes(s)\n", sizeof(x));
+	printf("Size of a long long int: %d bytes(s)\n", sizeof(y));
+	printf("Size of a float: %d bytes(s)\n", sizeof(f));
 	return (0);
 }
